backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
streamlit==1.41.1
openai==1.58.1
pdfplumber==0.11.4
openpyxl==3.1.5
Pillow==11.1.0
python-dotenv==1.0.1
python-docx==1.1.2
httpx==0.27.2
PyMuPDF==1.24.9
